+++ .gitignore
@@ -1,0 +1,9 @@
+# Folders
+__pycache__/
+
+
+# Files
+*.weights
+*.t7
+*.mp4
+*.avi
