--- YOLO3/__init__.py
@@ -1,6 +1,0 @@
-import sys
-sys.path.append("YOLO3")
-
-
-from .detector import YOLO3
-__all__ = ['YOLO3']
