+++ YOLOv3/__init__.py
@@ -1,0 +1,6 @@
+import sys
+sys.path.append("YOLOv3")
+
+
+from .detector import YOLOv3
+__all__ = ['YOLOv3']
