+++ deep_sort/__init__.py
@@ -1,0 +1,1 @@
+from .deep_sort import DeepSort
