+++ deep_sort/deep_sort.py
@@ -1,10 +1,13 @@
 import numpy as np
 
-from deep.feature_extractor import Extractor
-from sort.nn_matching import NearestNeighborDistanceMetric
-from sort.preprocessing import non_max_suppression
-from sort.detection import Detection
-from sort.tracker import Tracker
+from .deep.feature_extractor import Extractor
+from .sort.nn_matching import NearestNeighborDistanceMetric
+from .sort.preprocessing import non_max_suppression
+from .sort.detection import Detection
+from .sort.tracker import Tracker
+
+
+__all__ = ['DeepSort']
 
 
 class DeepSort(object):
@@ -70,6 +73,3 @@ class DeepSort(object):
         return features
 
 
-
-if __name__ == '__main__':
-    pass
